Move ISleepDurationStrategy into .Retry namespace
