+++ src/Polly.Shared/Polly.Shared.projitems
@@ -74,12 +74,12 @@
     <Compile Include="$(MSBuildThisFileDirectory)CircuitBreaker\AdvancedCircuitController.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Context.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Context.Dictionary.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Duration\ConstantBackoff.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Duration\DurationExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Duration\LinearBackoff.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Duration\ExponentialBackoff.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Duration\DecorrelatedJitterBackoff.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Duration\ISleepDurationStrategy.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Retry\ConstantBackoff.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Retry\DurationExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Retry\LinearBackoff.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Retry\ExponentialBackoff.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Retry\DecorrelatedJitterBackoff.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Retry\ISleepDurationStrategy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExceptionPredicate.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DelegateResult.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExecutionRejectedException.cs" />
